fix(rebuild): only clean up locks on failure

+++ scripts/rebuild.sh
@@ -11,8 +11,11 @@ source "$(dirname "${BASH_SOURCE[0]}")/helpers.sh"
 # Need this to avoid some wacky pre-commit hook issues related to if rebuild fails and
 # flake.lock stays staged, which ends up wiping out all changes due to stashing bug
 cleanup() {
-	git rm --cached -f flake.lock 2>/dev/null || true
-	rm flake.lock 2>/dev/null || true
+	if [ $? -ne 0 ]; then
+		red "Rebuild failed, cleaning up lock files"
+		git rm --cached -f flake.lock 2>/dev/null || true
+		rm flake.lock 2>/dev/null || true
+	fi
 }
 trap cleanup EXIT
